backend/requirements.txt
# Data Science & Math
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.11.0
scikit-learn>=1.4.0
umap-learn>=0.5.5
networkx==3.2.1

# API & Networking
httpx>=0.25.0
tenacity>=8.2.0
requests>=2.31.0
openai>=1.0.0

# Database & Environment
supabase>=2.3.0
python-dotenv>=1.0.0
python-louvain==0.16

# NLP
textblob>=0.17.1
